[Vulkan][Codegen] Spir-V codegen, correct labels/blocks in WhileNode. (#8013)
Previously, the WhileNode assumes that evaluating the loop condition
will not introduce any additional labels.  If this assumption is
violated, such as for a WhileNode whose condition is an if/else
statement, then the OpLoopMerge instruction appears in the wrong
block.

The unittest added exercises this code path, but doesn't yet trigger a
failure.  Once spvValidate is enabled for all vulkan codegen, then
this unit test will catch the failure mode.

Co-authored-by: Eric Lunderberg <elunderberg@octoml.ai>