[Vulkan][Refactor] Split out vulkan.cc into separate distinct functionality. (#8157)
This is in preparation for additional refactoring.  Functions are
organized according to group similar functionality together, to
minimize the amount of file-to-file transfers needed later.  The main
divisions are between VulkanDeviceAPI,
VulkanModuleNode/VulkanWrappedFunc, VulkanThreadEntry, and
VulkanContext.

Other than minimal renaming of private functions and addition of some
comments, this commit should have zero changes to the functions
definitions themselves, only to their arrangement within the
src/runtime/vulkan directory.

Co-authored-by: Eric Lunderberg <elunderberg@octoml.ai>