[Refactor] Remove dead code from depthwise_conv2d for Intel graphics (#8381)
After fix 65b19cc, I saw that it should be necessary to do the same fix
for depthwise_conv2d for intel graphics. I saw that we never used the
removed code and it is just the same code from
cuda/depthwise_conv2d.py. So we can use the cuda implementation when it
will be necessary.