[CMake] Split out libinfo.cc into a separate target. (#8520)
Every `*.o` file in the cmake-generated makefiles have a dependency on
the target's `flags.make` file.  The `flags.make` file contains the
compiler flags for all objects in a target, not just the `*.o` file
currently being compiled.  As a result, even though `libinfo.cc` is
the only file that has the `TVM_GIT_COMMIT_TIME` and
`TVM_GIT_COMMIT_HASH` definitions, every file in the `tvm_objs` target
was recompiled whenever the commit id was changed.

By splitting `libinfo.cc` out into a separate target, no other files
need to be recompiled when committing or changing branches, unless
there are actual changes to the file.

Co-authored-by: Eric Lunderberg <elunderberg@octoml.ai>