[microTVM][RVM] Fix base-box-tool command in README.md (#8613)
This commit fixes the platform argument order for base-box-tool.py
'test' command in the documentation about the RVM. Currently the example
in documentation places <platform> before option
[--test-device-serial=<serial>], whilst the correct order is after all
the options, so trying to use the 'test' command arguments in the order
as suggested by the documentation will not work.

This commit also fixes a typo (inovke -> invoke).

Finally it tweaks a bit the text format: lines with maximum 80 columns,
a better diagram format for the dir structure, and a better format for
the bash commands. A link is added too for easy access to the
"microTVM Reference VM tutorial" found in tutorials/micro directory. A
couple of command examples were also added to the documentation.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>