[Relay] Introduce Executor and Runtime representations with associated registries (#9246)
This is the underpinning Executor/Runtime objects for https://github.com/apache/tvm-rfcs/pull/29, this doesn't change the wiring as yet since that's a pretty big change in itself. Most of this patch is TVM boilerplate, which I've tried to minimize - there's maybe some future work to decide how to more easily define some of these things.