[docs][bug] Add redirects for moved pages (#9394)
The documentation refactor moved many pages that have outstanding
links from search engines and other sources. Because we don't
have direct access to the .htaccess file for the TVM docs webserver,
this patch automatically creates manual http redirects from
a list of tuples that provide the old page with the relative
link to the new page.