[microNPU] Fix incorrectly calculated stride when converting NHWC to NHCWB16 (#9560)
Fixes an issue that causes strides to be incorrectly calculated when the
number of channels in the input is less than 16 and involves a conversion
from NHWC to NHCWB16. This is due to TVM being 'too smart' when analyzing
generated TE and removing compute that is deemed unnecessary. Consequently,
strides over data are incorrectly calculated leading to an output
mismatch.

The PR uses a reduce sum operation to trick TE's data dependency
analyzer into looping over a whole block (16), rather than the number
of channels actually used (< 16). This causes the calculated strides to
be a multiple of 16 which is required for NHCWB16 format.

Change-Id: Ibf76a94a12cebf51fa716fcac1de932a271c4a6d