[Hexagon] Allow execution on target or simulator from HexagonLauncher (#10454)
Setting ANDROID_SERIAL_NUMBER=simulator will execute the tests on
simulator instead of a hardware device.

This patch also introduces an environment variable HEXAGON_RPC_LIB_DIR
to specify the location of the hexagon_api binaries. If unset, the
code will look for the binaries in the same way as before this patch.