[skip ci][ci] hotfix Jenkinsfile (#10492)
This was broken by #10456 since that did not have a case for ignoring the `main` and staging branches. This adds it at the Jenkins level so the further scripts won't even be run if not necessary. This is marked `skip ci` but since the changes all happen before lint this is being tested.

commit-id:1655b8da

Co-authored-by: driazati <driazati@users.noreply.github.com>