[ci] Remove commit check on ci skipping logic (#10537)
* [ci] Remove commit check on ci skipping logic

This makes it very hard to use an sometimes out of the submitter's control (e.g. when Jenkins decides to push a merge commit before running CI) for dubious benefit (the PR title is where people are looking after-the-fact anyways, so having it in the commit message doesn't make much sense). This removes the check for the commit message in order to make the process smoother.

commit-id:dbd18808

* Address comments

commit-id:ecd2be81

Co-authored-by: driazati <driazati@users.noreply.github.com>