TVMC: Add new text/relay frontend (#10941)
* TVMC: Add new text/relay frontend

This feature enables passing a textural representation of a relay module to the tvmc command line.

Example: `tvmc compile relay.txt --target c --runtime=crt --executor=aot --executor-aot-unpacked-api=1 --pass-config tir.disable_vectorize=1 -f mlf`

Currently it is not possible to supply parameters as it is mainly intended to be used for testing certain relay functions or operators. In the future (with minor changes to the tvmc frontend api) params could be passed via an additional i.e. `params.bin` file

This commit also adds minimal unit testing of the added feature.

Resolve PR comments

TVMC: add warning if relay frontend is used

* [TVMC] populate parameters with random values instead of ones

* [TVMC] Relay frontend: do not populate input tensor buffers if --input-shapes is provided

This prevents that the constants inputs are used for Constant folding,
thus changing the complexity of the model.

If there would be a way, to distinguish between model inputs and parameter this
workaround would not be required.

* [TVMC] Relay frontend: check provided file contents before calling tvm.parser.fromtext()