[TVMScript] TracedObject class that simplifies tracing ObjectPaths (#12299)
Motivation: when printing a piece of TIR, we need to track an ObjectPath from the root TIR object to the currently printed object. This means that we need a convenient way to maintain an ObjectPath whenever we access a sub-object, e.g. via an attribute.

Tracking issue: https://github.com/apache/tvm/issues/11912