[ACL] Adjust mobilenet test for Keras 2.9 (#12541)
In Keras 2.7, one "reshape" operator was removed from
the Mobilenet model, making our test which verifies the
number of operators to be incorrect.

This patch adjusts the operator count so that it is in line
with the changes in Keras. For reference, the change in
keras repo was done in hash b6abfaed132 "Remove unnecessary
reshape layer in MobileNet architecture".